Extract boost calculator helpers
